sw switch utility for fetching sdk version
Summary:
As titled, read the SDK version from the config if present. Return nullopt otherwise.

Hw test and link test configs will not have the SDK version set and hence its critical to set nullopt so that we fallback to default stream type.

Read the detailed summary in D49358544 for more details.

Reviewed By: msomasundaran

Differential Revision: D49358546

fbshipit-source-id: e77e7cc6f3b1e4e208452c17bc10166a16493479