set queue stream type based on the sdk verison
Summary:
Tajo changed the stream type from ALL to UNICAST with P4 WB SDK.

1.42.1  (previous prod) -> ALL
1.42.8 (current prod) -> ALL
1.65.0 (future prod, p4 WB) -> UNICAST

Following sequence fails today:

- Config is changed to set the stream type to use UNICAST for all queues (cpu, uplinks and downlink queue configs)
- When the switch inits, SwSwitch has created an instance of HwAsicTable which inturn creates an instance of HwAsic
- ApplyThriftConfig queries the HwAsic instance from the HwAsicTable for default stream type.
- For Tajo, the default stream type is set to ALL.
- There is a mismatch in what the default stream type vs whats being configured through the agent config
- ApplyThriftConfig resets all the queue configs and sets the stream type to ALL by using the default stream typ
- Agent cores

After having discussions with jasmeetbagga phshaikh and msomasundaran, we decided to go with passing the SDK version to the HW Asic. Hw Asic table already parses the config for switch info. Along with that, we can parse the config and set the SDK version on the Hw Asic.

This will help in two ways:

- Keep the code consistent from wedge agent and from the test perspective - Setting the SDK version to HW asic and leverage that to return default stream type.
- Any future changes which require the SDK version in the Hw Asic can leverage it.

If the config do not have the SDK version (HW Test or link test), fallback to use the default stream type. Hw Switch ensemble will set the right stream type based on the SDK version for HW Test.
If the SDK version >= 1.65.0, set the stream type to UNICAST. Else fallback to default stream type.

Reviewed By: jasmeetbagga, msomasundaran

Differential Revision: D49358544

fbshipit-source-id: 32d235e00e3091d26949f664c876f07e0b1d2a2f