For SFF, if the tx_dis/lpbk capability is present then make diags capability true
Summary:
The tcvrState's diagsCapability->diagnostics() support should be true when either Tx_disable, Rx_disable or Loopback is supported. This is the way it is in CMIS side. Putting the code to set the diagsCapability->diagnostics() to true when either of drag capability is there.

This is needed because in LinkTest class when we check the port for a feature then we look for diagnostics() field also

Reviewed By: harshitgulati18

Differential Revision:
D49464365

Privacy Context Container: L1125642

fbshipit-source-id: 66ff49333e723b3869bbd7716ddb8db71a2e6fa5