Fix HwTransceiverResetTest.verifyResetControl for CMIS optics
Summary:
When a CMIS optic is in reset, it reports UNKNOWN in the ModuleStatus register (byte 3, lower page 0). However, the test was assuming that i2c would fail. I2c only fails on SFF optics.

Fixing the test for CMIS optics in this diff by checking for moduleStatus.

Reviewed By: rajank7

Differential Revision:
D49530506

Privacy Context Container: L1125642

fbshipit-source-id: f390747a4cc9e871c633c35ae4a1e8a80ca4d1a9