Check tx disable capability before doing optics remediation test
Summary: The testOpticsRemediation is flaky sometimes. This test depends on tx_disable capability of optics. One improvement we can do to this test is to make sure the optics is capable of tx_disable before running the test. We need to do the test on the optics which have this capability.

Reviewed By: harshitgulati18

Differential Revision: D48830681

fbshipit-source-id: c0bc80f6f30b53e4ae75b10134db126d48424435