wedge_qsfp_util to use service based or direct acces for module type for test
Summary: Changed the wedge_qsfp_util code to either use common function to getModuleType. Earlier it used the direct_i2c access function and that caused many tx_disable command to fail because the i2c conflicts with qsfp_service. With this change we should not have any i2c conflict in test and the tests should pass

Reviewed By: harshitgulati18

Differential Revision: D49613650

fbshipit-source-id: 97d6b21281971be983ea65c475670b7e9fec0dcf