make the vlan header in the sflow packet conditional
Summary:
As titled, make the vlan header in the sflow packet conditional.

1.65.0 SDK and BCM SDK do not carry the vlan header when packets are punted to CPU whereas 1.42.8 carry the vlan header.

SflowMirrorTests.cpp is independent of SDK version and this case is very specific to the HW test. Hence ensure that the sflow header is always either equal to the original sflow packet received or 4 bytes more (vlan header)

Reviewed By: jasmeetbagga

Differential Revision: D49575814

fbshipit-source-id: b702ac12d751b6bad988c25f9f75c573b03d3353