Don't visit set values that have not changed
Summary:
I noticed for sets we sometimes emit a delta for values that haven't changed. I think because we use unordered_set, set_symmetric_difference will see changes for some keys even if it exists in both old and new set.

To fix, we can skip emitting a delta if the item exists in old and new

Differential Revision:
D48623126

Privacy Context Container: L1125642

fbshipit-source-id: bc4a58d67919a1c2892aa1351542454075b0faff