PatchBuilder skeleton
Summary:
Adding skeleton for build operation of Patches. Details on Patch in D47135229.

The idea here is to use DeltaVisitor in conjunction with a custom TraverseHelper. The TraverseHelper will be in charge of creating a Patch tree mirroring the delta traversal. The Traverser will keep a "path" of PatchNodes from root to the current node we're traversing. Then once we get an actual visitation at a modified node, we can set the value at the current node.

Differential Revision:
D48445189

Privacy Context Container: L1125642

fbshipit-source-id: 5df319c389b2c90679adbcde4c16e27e0d9f94b7