Template TraverseHelper push/pop on TC
Summary:
Templating these methods on type class of the pushed/popped path. We need this for implementing patch builder in coming diffs since we need to know the type of the node we're traversing.
Thrift type classes (specifically container types) are templated on the key/value type classes. So in the hope of reducing # of generated types here I've opted to introduce this ThriftSimpleTC enum to do away with those and reduce the number of combinations we need to generate for push/pop methods

I templated push/pop because I thought I might need to use static checks against the param (and because initially I was using actual thrift tc). So far I haven't needed to use the param at compile time. So at the top of the stack, based on feedback from priyankw, I think I'll change this back to a normal function param to simplify

Reviewed By: phshaikh

Differential Revision:
D48445188

Privacy Context Container: L1125642

fbshipit-source-id: c8146a233e0c811b6ee9c3a618d3f9705dc26220