Skeleton for PatchApplier
Summary:
Introducing a PatchApplier. This is in charge of the "apply" operation for Patches. see D47135229 for details on Patch.

The general idea here is simple to grasp. We traverse a Patch tree and the state tree in tandem, applying leaf patches (val / del) as we go and traversing further for patches that have children.

This visitor is a bit different in that it will just do the tree patching and not actually have a visit function (so it's technically not a visitor per-se).

I think alternatively we could have just added methods to Nodes to traverse the Patch tree, but I figured we can keep it seperate so clients don't need to include and build the logic unless they need it

Differential Revision:
D47135231

Privacy Context Container: L1125642

fbshipit-source-id: a340dc0c2f85ca2ad83ebfca68033ae770726c6a