Replace PmUnitPath vs SlotPath and DevicePath.
Summary:
PmUnitPath was a confusing notion.  It was pointing to both PmUnits
and devices in PmUnits, and was constructed with both slots and devices.
Instead we introduce two new definitions here. SlotPath which is entirely made
of slots, and can point to a location where a PmUnit can be plugged in.
DevicePath shows where exactly a device is plugged in. This also makes the code
more readable.

Differential Revision: D49711238

fbshipit-source-id: 24a0e616de0ca7c6f43ac4caa1d909917e307221