Retire ThriftStorage and usages
Summary:
This library was added and used before thrift_cow. Operation are comparatively it is much more inefficient since we're not dealing with cow trees. We havent seen a need to keep this around for over a year so going ahead with some clean up.
I wanted to remove the visitors as well but looks like there are still some usages (only in FsdbNormalizer afaict). So keeping them around, they're more useful than ThriftStorage anyways

Differential Revision: D48810804

fbshipit-source-id: e5f789060586559cce7917f69d5b56f263161b8b