Enable transaction switchstate updates in split agent
Summary: Enable transaction updates in split agent. This is based on sdk version info in config to determine whether the switch is SAI or non SAI. For mono, we are still using the indication from hw switch. This will be removed once all tests configs are updated to populate sdk version info

Reviewed By: jasmeetbagga

Differential Revision: D49601494

fbshipit-source-id: 6e5362465562af57e7bdcb0d26cbb3e983b16ef4