Support PatchBuilder for non-struct nodes
Summary: PatchBuilder assumed that the node passed in a struct. This will be the case for all publishers (patching their root which are all structs). But in general this is not true. Since we can have subscriptions at any path we absolutely need to support building Patches at every node

Differential Revision:
D48848689

Privacy Context Container: L1125642

fbshipit-source-id: 2f7d6268e24c8293b2bf2a8dd5dea4e08047b3c6