Dissallow patching immutable fields
Summary:
Currently this is not an issue because in set PatchApplier we handle all modifications and not recurse down. So we never end up in primitive PatchApplier. Now I've run into this because of next diff, trying to apply a patch at a path rather than root. By doing trying to do that I've technically allowed the primitive PatchApplier to be specialized with an immutable ThriftPrimitiveNode (only comes from sets), which will end up not compiling since we'll try to mutate in applyPatch.

Fix is simple, just turn the compile time error to a runtime error :D. In practice we'd never be trying something like this so shouldn't be an issue

Differential Revision:
D48832280

Privacy Context Container: L1125642

fbshipit-source-id: 3d8aa61034098195a4216cb54dbfbf2840ad9292