Symlink agent_hw_test to prod_invariant_tests for backward compat
Summary:
Next diffs will show that we the netcastle framework
is referencing this directory. So we can't really rename till
a new netcastle binary is rolled out. So to tide over this
situation, create a symlink in the interim. When netcastle
is rolled out we will retire this symlink

Reviewed By: shri-khare

Differential Revision:
D49978422

Privacy Context Container: L1125642

fbshipit-source-id: 48c7ad77d193f8ab3e535a36a6414288c5a8c49c