Program remote interface route to drop in hw
Summary:
In SW switch, remote interface routes are treated as directly connected routes, in order to aid ECMP resolution on remote system ports. However in the hw, such subnet route should be pointing to drop.

Adding a new flag of `isLocalRif` in router interface handler. When dealing with interface routes, sai switch will check the flag and point it to drop if not local.

Reviewed By: jasmeetbagga

Differential Revision:
D50026245

Privacy Context Container: L1125642

fbshipit-source-id: a504817a6ad64bb199c548297496551e9e1d2c76