Add default profile Id for a port if the FSDB update has not arrived yet
Summary: When we call the setExternalLedState() from a thrift and if the agent has not yet updated the switchState for the port then the LedManager will not know what is the port profile for that port. Due to this the subsequent function for setLed in HW will fail. To fix this, when setExternalLedState is called and if the Agent has not updated the SwitchState yet then we just throw from there. The LED service is not assumed to be operational for a port till the update arrives from Agent (FSDB) for that port

Reviewed By: harshitgulati18

Differential Revision:
D49900936

Privacy Context Container: L1125642

fbshipit-source-id: 124367e3eab3ea3b6deac7ccab3554d05619be95