Skip programming iphy for tcvr with no ports
Summary:
Blanco bypass modules have no ports associated with them, so programInternalPhyPorts will throw an exception, and creates misleading error logs. We should skip trying to program the iphy if there are no ports to program.

Originally I planned to make this change in the qsfp service (skip calling the thrift API if we find a blanco module), but I found that we still call this function even if no module is present  (https://fburl.com/code/qr8pdxvl), so we still need to handle this in the thrift handler to catch all cases where this function is called for a bypass module

Reviewed By: rajank7

Differential Revision:
D50205104

Privacy Context Container: L1125642

fbshipit-source-id: 895b57179f9c806bd44d086f0489ece3aa6e3ecc