Fix port reusage in HW test
Summary:
HW tests fails when data corral service is already running. This has been causing node failures in conveyor.

This looks to be an issue for some time. Basically, the test uses the same port, so ThriftServer throws an exception:

````
E1016 11:49:45.689487 3589628 ThriftServer.cpp:654] Got an exception while setting up the server: failed to bind to async server socket: [::1]:5971: Address already in use
unknown file: Failure
C++ exception with description "failed to bind to async server socket: [::1]:5971: Address already in use" thrown in the test body.
````

This was going to be fixed as part of D50134925 but it's blocked on LED verification and this issue is blocking platform_stack push so putting it out as a separate diff

Reviewed By: somasun

Differential Revision: D50333929

fbshipit-source-id: 5c3e40749c0be8a1c0baa8b48346602a36860383