Introduce isSai() method in mono and split agent ensemble
Summary:
This is towards refactoring some of the sai/bcm specific util methods to the non-hw dependent layer. In order to build sw agent test binary with no hw dependency, agent ensemble needs to understand whether it is running with sai or non-sai.

For mono agent ensemble, agent initializer checks if the platform is sai or non-sai. For split, sw switch has no understanding of the SDK and its version. After the conversation with msomasundaran, we decided to have a command flag to pass in for test binary to understand sai vs non-sai.

Reviewed By: msomasundaran

Differential Revision:
D50415170

Privacy Context Container: L1125642

fbshipit-source-id: a21d1305151d72211e738f6595cb0030ef109534