Create a instance-id base for each pci-device found
Summary:
We need to use the same instance id when we try to create the same i2c
controller again (for debuggability, traceability). We can ensure this by
keeping an offset for each fpga discovered in the platform, and using that as
the base instance-id for all devices created within the fpga.

Differential Revision: D50561688

fbshipit-source-id: 11a07cfc3082560c8456fae288cc824e6baacc97