Mark SystemPort enabled field as deprecated
Summary:
Towards deprecating system port enabled field.
Setting this to anything but true is not accepted
by sai adaptor. Having this field here makes it
a bit tempting to mirror the local ports admin
state here. But copying this from a disabled port
will lead to a crash.

Stacked diffs to percolate this change

Reviewed By: shri-khare

Differential Revision: D50633023

fbshipit-source-id: cb0c9da7a941567773318b5a7665bffbf8c864d0