Kill SystemPort::getEnabled API
Summary:
enabled field is deprecated for sys ports

Context:
Towards deprecating system port enabled field.
Setting this to anything but true is not accepted
by sai adaptor. Having this field here makes it
a bit tempting to mirror the local ports admin
state here. But copying this from a disabled port
will lead to a crash.

Reviewed By: shri-khare

Differential Revision: D50634626

fbshipit-source-id: e27fffff1d8a9b6be387a980383bc0d089bd3c0e