Fix HwQsfpEnsemble::didWarmBoot
Summary:
We were returning warmBoot_ from this function before but this variable never gets updated anywhere so didWarmBoot was always returning false in the tests.
Fixing this by reusing canWarmBoot function from WedgeManager/TransceiverManager.

Also made this a protected function as I plan to use it in the tests subclass later in this stack

Reviewed By: rajank7

Differential Revision:
D50609829

Privacy Context Container: L1125642

fbshipit-source-id: 58cde9206558b91129e0fdc65d11c75a00a5cdb9