Sync stats from hwagent after reaching configured state
Summary: Execute stats collection and sync logic in hwagent after hwswitch transitions to configured state.  The status update is done from function scheduler outside hwswitch. The function scheduler is created at init time and the run state is checked when updating stats. Also noticed that the hwswitch was transitioning directly to configured state. Added an initialized state transition in between to keep it in sync with mono.

Reviewed By: jasmeetbagga

Differential Revision: D50713059

fbshipit-source-id: c7154ffa06df9fcf8eb9b12c34259c48db038b3e