weutil : Define the new FbossEepromParser class (header file)
Summary: This FbossEepromParser class is the class that will be shared with OpenBMC Code base. This class has all the logic for reading and parsing EEPROM, but does not depend on folly or thrift so that it will also compile in BMC;  another class, FbossEepromLib will have all the fbcode specific folly and thrift related logic. This class will be defined in the later diff of this stack.

Reviewed By: somasun

Differential Revision: D49491158

fbshipit-source-id: 165448442c249f0b8f192358d9b25a53a975b3e1