Start reading FEC corrected bits stat
Summary:
SAI_FEC_CORRECTED_BITS stat is added in SAI version 1.13.
In this diff, I add the code to read this stat periodically (along with the other FEC stats) and store it in HwPortStats. In the next diff, this stat will be used in the Pre-FEC BER calculation for PhyInfo struct.

Reviewed By: zechengh09

Differential Revision: D50650367

fbshipit-source-id: 69e5afe230dd1febc7dcb7d2ec52de92ddefe7c1