Fix systemCores computation during switch create
Summary:
The current computation assumes continuous switchIds starting at 0, and then as
it loops over dsfNode map, it maintains a running sum of systemCores, which it
uses during switch create.

However, this logic fails if the dsfNode map has switchIds does not start with
0, as then systemCores is not correctly updated.

Fix it by deriving systemCores for a given dsfNode from its switchId.

Reviewed By: jasmeetbagga

Differential Revision: D50766544

fbshipit-source-id: 8160679828a155331c0c9641fc3476fbfd03d401