Fix removeRemoteInterface helper api
Summary:
We removed the RIF from newState, return
that, not the curState. The previous version would
continue to work as long as we passed in a unpublished
state, since then newState and remoteState would point
to the same switchState. But if the input state was
published, we would not get the same result.

Reviewed By: shri-khare, nivinl

Differential Revision: D50871715

fbshipit-source-id: b186086d77a13a23ef9bdf72ae8cf13ed9f0476f