allow graceful service loop termination in ReconnectingThriftClient
Summary:
ReconnectingThriftClient implements cancel() API by canceling the service loop as part of canceling the Thrift client.

As part of publisher GR support, publisher client would send a disconnectReason indicating GR  to the server, and then close the service loop.
So we need a mechanism for the Thrift client to wait for service loop to complete before completing Thrift client cancellation.

This change adds hooks in ReconnectingThriftClient to indicate graceful service loop cancellation, and invoking a callback invoked when service loop completes gracefully.
The callback and override to setState() in the base class will be used in next diffs to implement sending disconnectReason to server.

Reviewed By: peygar

Differential Revision:
D50134966

Privacy Context Container: L1125642

fbshipit-source-id: 3e66ed151fb4edf3bd5b58aad6addd1b6e7d5a12