optionally check and enforce subscriber permissions
Summary:
Add flags checkSubscriberConfig and enforceSubscriberConfig to respectively check and enforce per-subscriber permissions.
These flags are used to check and lock down use of extended subscriptions.

At this time, by default we only check the config but not enforce it.
A counter is incremented if any subscription request fails the permission check. We will monitor this counter to catch if there are any unknown subscribers using extended subscriptions.
In a later push enforcement of extended subscription permission will be enabled.

Reviewed By: peygar

Differential Revision:
D50783589

Privacy Context Container: L1125642

fbshipit-source-id: e64419e9bb1b8d24c0db128c6401dabdf1ddb7b9