API to get ire error count
Summary: This diff introduces a new API to get the ire error count in the HwSwitchFb303Stats class. The code changes include adding a new method to the HwSwitchFb303Stats class to get the cumulative value of the ireErrors_ counter. The changes also include updating the HwSwitchFb303Stats.h header file to include the new method.

Reviewed By: nivinl

Differential Revision: D51060128

fbshipit-source-id: d5b364331eeebb5f66cf8353e7cee1595ba4d19e