Fix the broken led_service unit tests
Summary:
The led_service unit tests had got broken earlier because sometime back we put the config file and the FSDB subscription happens only if the config variable isLedControlledThroughService is set as True. This was not the case in UT so it started failing.

This diff fixes it by creating a test config and passing this config file to LedManager while init.

Reviewed By: harshitgulati18

Differential Revision:
D51079549

Privacy Context Container: L1125642

fbshipit-source-id: c005cdd7c42879d57e79e346b20b6d958b1df6ef