Get ProfileID matching speed and lanes when constructing initial SwPort after cold boot
Summary:
Reusing the profiles returned by findPortIDAndProfiles instead of another getProfileIDBySpeed which only returns the profile matching a given speed.

Using getProfileIDBySpeed is buggy since it will not take into account the lanes along with the speed. We could have multiple profiles with the same speed but different number of lanes (for example 400G-4, 400G-8). Using the result of findPortIDAndProfiles guarantees improves the situation by matching both speed and lanes.

It is still possible that this is not the final profile as we also need to account for MediaType here. However, the SDK could initialize ports with any media type right after cold boot and we may not have a profile matching that. Thus, we pick 'A' profile matching speed and lanes, and later update the profile when applyThriftConfig comes in

Reviewed By: zechengh09

Differential Revision:
D51538668

Privacy Context Container: L1125642

fbshipit-source-id: 9f88014c4a0d6035e14028db526d25f29c7ad64e