Unify the APIs used in validatePGQueueChanges test
Summary:
`checkSwHwPgCfgMatch` is already making sure that we check the content of each
PG b/w HW and SW. This along with a size comparison of PGs will ensure that the
functionality provided by checkSwHwPgIdSetMatch() is covered.

Remove the new API added for PG ID validation and incorporate the changes in the
existing API and make sure that the PG is programmed with headroom so that its
considered lossless.

Marking `validatePGQueueChanges` as `unsupported` for Jericho in D51533481.

Reviewed By: jasmeetbagga

Differential Revision: D51530066

fbshipit-source-id: 615c45ea311e07b1f6fd7dd7e64c36b10866f340