Kill DECLARE(enable_state_oper_delta)
Summary:
This flag is no longer used in the codebase, so
remove the forward declaration.
Next diff will remove the flag completely. I can't do that
in this diff as some prod configs still have the flag and
we need to wait for config generation past D51643763 landing.

Reviewed By: shri-khare

Differential Revision:
D51644607

Privacy Context Container: L1125642

fbshipit-source-id: f31a4932b55d1a7e460d560e9f2a55b5dc8ec442