Add support for qsfp_reset to go through service
Summary:
Adding a command to the wedge_qsfp_util (--qsfp_reset) to go through qsfp_service if service is up, or reset directly via CPLD/FPGA when the service is down and the --direct_i2c option is added.

This is in case the qsfp_service is up, it has to know about a reset, to make the service through proper statemachine changes.

e.g.
./wedge_qsfp_util eth1/1/1 --qsfp_reset

QSFP service will still do the same reset with the old command: --qsfp_hard_reset whether the qsfp_service is up or down. This will be deprecated once qsfp_service changes in this diff are in the fleet.

vblake and others CCd on diff based on usage in these files:

https://www.internalfb.com/code/search?q=repo%3Aall%20qsfp_hard_reset&lang_filter=python

Reviewed By: harshitgulati18

Differential Revision: D51479182

fbshipit-source-id: 6b59ed7f3a7066a757e98dd92562325fb61847cd