Add a method to get map for sys_port_id -> port_id->port_summary
Summary:
As titled. This function will be used in next
diff to populate port association for system port
interfaces in CLI commands. Till now we were handling
only vlan interfaces.
Extending old clis as I could not find a equivalent
command in fboss2 cli. Further interface on fboss2 CLI
is associate with a port and not a l3 intf

Reviewed By: harshitgulati18

Differential Revision: D51739305

fbshipit-source-id: 43cfb00ab18e543afdbc97da2e8e5342d3d85cc8