Refactor port map init code
Summary:
This code is currently part of loadConfig because previously we used to rely on agent config to get the ports. Last half, we removed the agent config dependency and instead started looking at ports from platform mapping. Thus, this code can be moved out of loadConfig into the ctor.

The other advantage is that wedge_qsfp_util can now access portNameToModule_ too as in the next diff

Reviewed By: rajank7

Differential Revision:
D51793942

Privacy Context Container: L1125642

fbshipit-source-id: f7c6a670a10f95774a3b60bb854b9ae5ca732a85