Fix ODSCounter hw test failure
Summary:
This has been causing hw test node failure.

It was because we never controlled fan in the test case so ODS counter is empty. But for some reason, when the local binary is ran on a switch, all test passes. I suppose ODS counters were still available from other tests

Reviewed By: somasun

Differential Revision: D51812982

fbshipit-source-id: f663df07cbeaa837eb47bf88c2a749c09ab6aa50