Change the font color in the "Status" field of "fboss2 show transceiver" CLI output
Summary:
"fboss2 show transceiver" - change the font color of the "Status" cell in the output of this command so that it becomes easier for a user to spot a bad port from the output.

Create a new function to convert status boolean to string with a representation of that string with a color - green for Up and red for Down status.

Reviewed By: harshitgulati18

Differential Revision: D51812706

fbshipit-source-id: c4222bd150747220dc08c99179d68e96e017ae7a