Update Ingress interface if present
Summary:
If we follow this sequence:
1. Warmboot upgrade with "sidelinks" config (new vlan 3000 and l3 interface)
2. Warmboot downgrade after removing "sidelinks" config.
3. Warmboot upgrade again with "sidelinks" config.

it asserts on bcm_l3_ingress_create in programIngressIfNeeded. The reason is that l3_ingress interface is present and we did not have BCM_L3_INGRESS_REPLACE set.

We were setting replace flag as false if vlanMac2Intf is not present in warmboot cache and assuming that l3_ingress interface is also not there. However due to a bug which is fixed in D51830932 we have a situation in where we have l3_ingress  interface present in the asic.

This diff is a temporary diff to be able to warmboot enable "sidelinks" feature in RTSWs where we had hit the bug.

Reviewed By: daiwei1983

Differential Revision: D51794128

fbshipit-source-id: e0f9751c6b942e5943e5d363909009eb71f43901