Use gflag for platform mapping profile.
Summary:
Add a gflag to decide which platform profile to use. For backward compatibility for now, check for gflag being set to "inference" and if not, fallback to the netwhoami way.

We want to phase out the netwhoami functions eventually and rely on cfgrs only, but need to handle plan for the qsfp_wedge_util in eventual diffs since it also leverages netwhoami and will get broken as it can't rely on the cfgr gflag.

Plan is to move netwhoami based checks to as separate lib invoked only for qsfp_util where all rack types will be explicitly added.

Reviewed By: harshitgulati18

Differential Revision:
D51964343

Privacy Context Container: L1158612

fbshipit-source-id: dbd013537a5260c27f23bc00dd80bcabc7a3094a