Make trapPktsOnPort test actually trap pkts due to pkt trap ACL
Summary:
Currently in the test, no nbr entry was setup, so
that meant that the pkts would trap to the ASIC via IP2Me rules.
This was not the intention of the test. The test was meant to
test our ACL based pkt trapping functionality.
Fix this by programming neighbor entry

Reviewed By: srikrishnagopu

Differential Revision: D52047923

fbshipit-source-id: e0c8527afa34e79cfa897675fb42ab8bbe028baa