Feature for Link Active/Inactive based Unisolate/Isolate
Summary:
Some ASICs require min number of links to be active i.e. up + bi-directional
connectivity is establised, before the ASIC can participate in the network.

Provide a feature flag for it.
Only enabled for J2 at the momemt, in future, we will validate and enable this
for J3 as well.

Reviewed By: jasmeetbagga

Differential Revision:
D52019013

Privacy Context Container: L1125642

fbshipit-source-id: f5fd2c8ff0b3b28f3bdbeaa944fa462e98eb0c12