TX Ready cb: query/process in the same context that registers cb
Summary:
Some ASICs require min number of links to be active i.e. up + bi-directional
connectivity is establised, before the ASIC can participate in the network.
This diff enhances the agent code to dynamically monitor as link turns
active/inactive and unisolate/isolate as needed.

This diff: as titled.
See comments in the code.

Reviewed By: jasmeetbagga

Differential Revision: D52041719

fbshipit-source-id: c8ecc7d4d6927c006b6628806a5e450a510cf64a