fixing logic of verifying core in agent wrapper test
Summary:
a core dump directory name depends on a thread which receives SIGABRT and terminates the process, it is not necessary to begin with name of a process.

as a result fixing the logic to verify core in agent wrapper test.

a core directory name is guaranteed to have pid and process name, so checking if core-directory is created and if it has core and metadata file which is created by core-copier

Reviewed By: jasmeetbagga

Differential Revision:
D52141818

Privacy Context Container: L1125642

fbshipit-source-id: 1b18100086c7d68cec7e3bead5f74c36cecbd908