Start reading FEC codeword stats
Summary:
Fec codewords stats can be used to determine how a link is performing. If there are codewords with lesser number on symbol errors, then its a sign of a healthy link. If there are more codewords with larger number of symbol errors, its a sign of a degrading link.
Facebook
This feature is first supported on TH4 in 10.2.
CS00012302575

Reviewed By: zechengh09

Differential Revision:
D51881502

Privacy Context Container: L1125642

fbshipit-source-id: 15e59ecb96d944fec97b4ebca1fc5447eecc683d