Delete the tam objects and recreate to avoid multiple tam objects
Summary:
To address SMS out of bank error seen in the fleet, TAJO has a new notification which will notify FBOSS agent
on SMS out of bank errors in the fleet via SAI_SWITCH_EVENT_TYPE_LACK_OF_RESOURCES. However, this
would need the SwitchEventType to be modified and TAJO has a limitation that only one TAM object can be
created at any time. This means, warm boot upgrade is not possible to a new version which will try to create a
new TAM object as that would end up in 2 TAM objects being created. The work around is to delete the TAM
object and related objects during a warm boot and then create a new SwitchEventType which has the needed
notifications and then a new TAM object.

The above need to be split as 2 separate diffs, the first one just deleting the TAM objects during every WB and
once it rolls out, make the changes to introduce SAI_SWITCH_EVENT_TYPE_LACK_OF_RESOURCES in the list of
events, this is to ensure that round  trip tests can pass throughout.

Reviewed By: phshaikh

Differential Revision: D52190051

fbshipit-source-id: fb8ac32f520571e250272f1678986f739db2a78a