state delta comparison policy
Summary:
the diff introduces a delta comparison policy, which comes in two variants
- shallow : where contents of node are not investigated
- deep: where contents of node are investigated

delta comparison policy by default is shallow which means behavior as usual and simple existence of pointers imply difference.

RAII is used to control this policy, and can be created during stateDelta processing

```
{
  DeltaComparison::PolicyRAII policy(DeltaComparison::Policy::DEEP)
  stateChanged(StateDelta())
}
```

Note that RAII is not thread safe, and expectation is delta processing happens in a context of single thread and RAII is employed in that thread only.

Deep comparison is needed only for "resync" operation when an oper delta of full state may be sent by SwSwitch and may not have objects changed.

Reviewed By: jasmeetbagga

Differential Revision:
D52186695

Privacy Context Container: L1125642

fbshipit-source-id: b72914d5837b4ef572b38868067df02dded4844e