Create WedgeQsfp impls at init time
Summary: We seem to be creating them in updateTransceiverMap which happens in every refresh cycle. Since we need to create as many WedgeQsfp as number of transceiver modules supported on the platform and that is a static number, we should only create them once during init (this diff) and use those objects everytime (next diff).

Reviewed By: xanabani

Differential Revision: D52449414

fbshipit-source-id: 25551fe47a4cdd3e3806768ac697270851582400