Remove deviceType as we are no longer using this field
Summary:
Instead of completely removing the field, this diff replaces it with
isGpioChip boolean field.  We need to know whether a i2cdevice is a gpiochip,
to further find the gpiochip number for the gpiochip-i2cdevice. The alternate
approach could have been to introduce an extra value of "GPIOCHIP" to
deviceType (in addition to "SENSOR" and "EEPROM").  But we don't use the values
of "SENSOR" or "EEPROM" in deviceType anywhere.

Having said that, we might revisit this and introduce deviceType, if we have to
introduce another isX field in I2CDeviceConfig.

Reviewed By: xanabani

Differential Revision: D52633322

fbshipit-source-id: 1a5bf3de8850ca52be0067ebc63bb51d8bb59b56