Do not send initial cold boot state to HwSwitch
Summary:
The cold boot state constructed in SwAgent has only alpm state while cold boot state in HwAgent contains port and cpu port entries. The cold boot state sent from SwSwitch cannot be used to construct a full state in HwAgent as it will remove the entries in HwAgent. This is currently not a problem since we apply the oper delta to existing state in HwSwitch and don't try to migrate HwSwitch to SwSwitch's state. This change is in preparation to allow full sync of switchstate in HwAgent.

Notes:
- We still sent the initial state in warm boot. This is because we set the current state to initial state in SwSwitch so that config apply can work as it exists today. This also means that notifyStateObservers should be called on initial state in warmboot case.

TODO:
- full sync handling in HwAgent
- Similar changes for mono

Reviewed By: jasmeetbagga

Differential Revision: D52433191

fbshipit-source-id: f37678ee1f573b29d1b0dcc4c302af0fa4ea982a