Add a timeout for oper delta programming ack from hwswitch
Summary:
If a hwswitch is deadlocked, it can result in state update thread getting blocked for all switches since we wait for a response from all hwswitches when processing state update. Instead if one hwswitch is not responding, the oper delta command ack from the hwswitch will timeout after 10 minutes so that we can 1) identify stuck hwswitches 2) continue operating with remaining switches.

TODO:
increment a counter and export it

Reviewed By: jasmeetbagga

Differential Revision: D52492928

fbshipit-source-id: 9cc03e6ef909128f26e5e826d72a014b849abcac