scope resolver method to get Matcher for a given switch type and method to remove some switch ids from matcher
Summary:
This diff introduces changes to the HwSwitchMatcher and SwitchIdScopeResolver classes in the FBOSS agent codebase. The HwSwitchMatcher class now includes a new method called exclude() which allows the exclusion of switch IDs from the matcher. The SwitchIdScopeResolver class has also been updated to include a new method called scope() which returns a HwSwitchMatcher object for a given switch type. The changes in the SwitchIdScopeResolver class allow for the creation of a HwSwitchMatcher object that only includes switch IDs of a specific switch type.

The change will be useful in identifying scope for an isolated system port, a system port which is on another ASIC

Reviewed By: jasmeetbagga

Differential Revision:
D52611479

Privacy Context Container: L1125642

fbshipit-source-id: 75b73c34e22116f35c86f920795bc3181e375a4b