Get rid of gratuitous const on return by value
Summary:
Return by value is returning a copy not a reference
to internal object data. Const is gratuitous here, it gives
the impression of a missed const ref return, which is
not the intention here. If the caller wants to annotate
the returned value as const, they can store it in a
local const var

Reviewed By: srikrishnagopu

Differential Revision: D52720334

fbshipit-source-id: ed5ac5a17dee46fd329ab79c9a70c6d652e4513c