Always honor noHostRoute value passed in
Summary:
As titled. Previous diff had a bug in
that if we want from noHostRoute=true to noHostRoute=false,
we would just skip updating the attribute.
If it turns out that some platforms don't support this
attribute, we will change to passing in optional here.

Reviewed By: zechengh09

Differential Revision:
D52759238

Privacy Context Container: L1125642

fbshipit-source-id: 9d69503bafc0b86e342c3a8ab1d54f97ca76919f