Add counter for TTL1, hopLimit=1 pkt getting trapped to
Summary:
TTL or hopLimit 1 pkts are not necessarily a
error. A neighbor one hop away may intentionally set this
field to 1 to ensure pkts never go beyond one hop. This
is a common practice in e-bgp with directly connected
neighbors.
However, in DSF, with FSDB sessions over loopback IPs
we were not setting TTL/HopLimit to 1. Yet due to a
bug in programming, pkts were looping over the
recycle port and only then getting trapped to use
after TTL went down to 1. Add a counter to catch
such issues

Reviewed By: srikrishnagopu

Differential Revision: D52760277

fbshipit-source-id: 2b269bad31fea1794b96160c9f18f423b6ea51ac